Adding EXP and Projectile speed card. Improving card selection.

+++ include/render.hpp
@@ -17,7 +17,7 @@ extern SDL_Renderer* g_renderer;
 void RenderProjectiles(SDL_Rect camera);
 void UpdateRenderStopwatchWave(int& start_time, int& time_left, int screen_width, int& elapsed_time, int& wave, int& current_game_state, 
                                 TTF_Font* font, int total_pause_duration, DynamicText *stopwatch_text);
-void RenderExpBar(int screen_width, int exp);
+void RenderExpBar(int screen_width, float exp);
 void DrawThickRect(SDL_Renderer* renderer, SDL_Rect* rect, int thickness);
 void RenderCardSelection(int card_selected, TTF_Font* small_font, int screen_width, int screen_height, int level_to_update);
 void RenderHeader(int& start_time, int& time_left, int screen_width, int& elapsed_time, int& wave, int& current_game_state,
